prjfs: keep mount alive during notification handling
Summary:
When unmount is called, the mount is kept alive until all the pending
ProjectedFS callbacks have completed (D45135531) to avoid use-after-free.
However, the notification callback is handled entirely in the background and
would thus not prevent the mount from being unmounted, and not being freed.

In order to fix this, we need to keep the PrjfsChannelInner alive for the
duration of the notifcation handling, which as a side effect will keep the
mount alive.

Reviewed By: mshroyer

Differential Revision: D45251808

fbshipit-source-id: e78afbc388cdf75f6d8d4e774bced8fdf841db5e