server: add events for import and export stacks
Summary:
Expose `debugimportstack` and `debugexportstack` commands via the isl-server
WebSocket protocol so they can be used by the client.

Right now, the `debugimportstack` (write) command is not queued as an
operation. When editing a stack with explicit commit hashes, it does not want
to wait for other operations, since the stack is being edited as-is, editing
the rebased/amended stack is incorrect.

Reviewed By: evangrayk

Differential Revision: D45088038

fbshipit-source-id: 60584125136fcf2a4e15b0edc4ac892634e62f31