shared: make LRU friendly for immutable data structures
Summary:
I want to use a LRU cache where keys are tuples of immutable values
for linelog. But that's not possible with a native `Map` that uses
`Object.is` to compare keys. Change the LRU implementation to use
immutable map so key comparison is done via the desired `Immutable.is`.

Reviewed By: evangrayk

Differential Revision: D45154108

fbshipit-source-id: 2adc353a95cd4e3920f16285c887a9fd6534b1e1