animate: minimal container component for reorder animation
Summary:
I have done some exploration in the reorder/animation/drag-n-drop area and
concluded it's probably best to just DIY our needs. Summary of my learning:

Animation:
- The best API to use is Web Animation APIs. It's fast, flexible, supported
  by major browsers for years, and does not churn the `style` attributes.
  See https://mozdevs.github.io/Animation-examples/
- Only `opacity` and `transform` changes are fast. They trigger "composite",
  no slow "layout" or "paint".
  See https://aerotwist.com/blog/flip-your-animations/ ("FLIP") and
  https://huijing.github.io/slides/03-css-triggers/
- Most popular animation libraries use `requestAnimationFrame` and changes
  the `style` of DOM elements in Javascript. They follow the "FLIP" best
  practice, but somehow don't use the great Web Animation APIs.
- The actual work to implement animations with React is not that difficult
  as I initially thought.

Reordering:
- Can be done by animating the x, y offsets.
- If we ignore the "insert" and "delete" animations, the code is much simpler
  (not tracking `children` in `state`). Otherwise it'll somewhat re-invent
  `react-transition-group` internals.
- `react-transition-group` can probably be used together if we want "insert"
  and "delete" animations.

Drag-and-drop:
- Pointer events cover both mouse and touch. They are nice to use. We don't
  have to use drag and drop events.
- Does NOT have to be coupled with reordering animation. For example, we can
  just "render" the "preview reordered" items and send them to the reordering
  component. The reordering component gets the preview and does not need to
  maintain an extra "preview" state, which can be quite annoying.

This diff implements the reorder animation support. Drag-and-drop will be done
next.

ChatGPT contributed a lot of the initial implementation. I like its use of the
`React.FC` type to simplify the code a bit. I provided the concrete requirements
and did some clean up. ChatGPT used `ResizeObserver` to observe bounding box
changes but that's not enough (ResizeObserver does not observe offset changes)
and can be tricky to know if the states are using the old children or new
children.

What we actually want is the "old" states just before the animation
(useLayoutEffect). I added a side effect that runs before calculating the
rendered elements to do that.

Reviewed By: evangrayk

Differential Revision: D45405113

fbshipit-source-id: f57d99edc95237bcef099cf2f8636059885d69a5