perf: cache LineLog.recordText
Summary:
This is an easy way to reduce the cost of rebuilding linelogs after reordering.
This makes reordering a long stack feel relatively fast (reorder: ~60ms, then
renderRootSync in ~200ms).

An alternative fix would be making the commit stack smarter when rebuilding
file stacks and reuse existing file stacks if possible. But this easy fix seems
a good start.

Reviewed By: evangrayk

Differential Revision: D45495601

fbshipit-source-id: dfcc8976ffae5c54f2aada31368d6fd3fadc4c2a