types: use full hash internally
Summary:
Change the internal representation of a commit hash to use full hash.
This unifies the `Hash` type used by ISL and `ExportStack` and make them
comparable.

For UI display we truncate the full hash so they look the same.

For clipboard copy the full hash is used, which is consistent with GitHub and
CodeHub.

Reviewed By: evangrayk

Differential Revision: D45088033

fbshipit-source-id: 82a5e9de6d4f44e0056ea3a5aab8f3e66893d0c2