Implement chunking support in core sharding framework
Summary: With all the building blocks in place, this diff implements chunking support in the core sharding framework. Because the framework level contract is being modified, it affects all the services and jobs using sharding. In usual cases the changes are minimal but in x-repo and chunked jobs, there are some additional modifications. Note that these changes are not backward compatible so I will be stopping the existing chunked jobs (`AliasVerifyProd` and `AliasVerifyBackup`) before rolling out these changes. I plan to test these changes on the test sharding job.

Reviewed By: mitrandir77

Differential Revision: D45479019

fbshipit-source-id: 6e2290de356642a2dc33e35b56c9ab25c2571769