perf: immutable extension to avoid repetitive immutable.deepEqual
Summary:
When `immutable.is` returns `true` in O(N) time by calling `deepEqual`s, it
could be a disaster. The same `deepEqual` might be called again, and it's O(N)
again... The program is "stuck" in a slow state and is hard to recover from
it.

It is even worse with `<StackEditCommit />`. The component uses
`calculateDepMap` to figure out what operations can be done. `calculateDepMap`
is `cached` on the `CommitStackState`. The `CommitStackState` has O(N)
`stack`. The `cached` LRU cache needs O(N) deepEqual to `get` from the cache.
And we have N commits to render. So the deepEqual effectively becomes O(N^2)
slowdown and cannot recover from it. Practically that causes multi-second
browser freeze for reordering a 58-commit stack.

This diff introduces a container type `SelfUpdate` that updates its internal
value object (`inner`) to `other.inner` when `this.equals(other)` returns true.
So the next time `this.equals(other)` can use object reference comparsion and
be fast.

Now the question is should `x` be updated to `y` or `y` updated to `x`? What if
there is another `z` with the same value? Since our LRU caches are "global", we
ask objects outside the LRU cache to match the LRU cache references. Regardless
of how many versions of the same value, as long as those versions compare to
the LRU cache, they will be updated to a single object reference.

Reviewed By: evangrayk

Differential Revision: D45495600

fbshipit-source-id: b3b3dfc912b57fdc6cd2f3dbc3ee5b8004bbfce5