stack: add edit stack button per linear stack
Summary:
Add logic to draw an edit stack button for linear stacks. To reduce visual
noise, similar to the goto button, the button only shows up when hovering over
the stack.

The button does not yet do anything.

The button disables itself when:
- The stack is not linear. I don't want to deal with the complexity right now.
- The stack has pending changes. Editing might cause conflict or overwrite
  files undesirably.
- The stack has obsoleted commits. Editing would cause divergent commits
  (multiple "alive" commits with the same commit message which are usually
  confusing).

I added some tooltip to explain why the button is disabled.

It would be nice if the tooltip auto hides when the element transits to
`opacity: 0`.

Reviewed By: evangrayk

Differential Revision: D45088040

fbshipit-source-id: 30c4787f1fe40b632dd660256c785781349bef14