progress: hide `hg debugimportstack` args and output
Summary:
`debug` commands are considered private and its interface is unstable. We want
to reserve the right to change them without worrying about compatibility
breakage.

It's also meaningless for users to run `debugimportstack` without knowing what
the stdin is. The output is not for human consumption either.

Therefore hide the command and output.

Reviewed By: evangrayk

Differential Revision: D45571179

fbshipit-source-id: c838f2169eee5b4a991d5d4ba31125b3c13b902a