debuglocks: deprecate ability to force lock release
Summary: The force release behavior is broken, and may no longer be necessary. It is broken because it unlocks by deleting the legacy compatibility lock file, which is not load bearing for locking anymore. It could be updated to delete the right lock files, but I don't think that is necessary since locks on all platforms are released automatically when the locking process exits.

Reviewed By: quark-zju

Differential Revision: D41998655

fbshipit-source-id: f3833e07668d108b7a2464dd272a03c76379f32b