lock: kill python "releaselock" function
Summary: This was only used by the "debuglocks" command. Instead, make it get the lock information from the LockHeld exception. This is a better approach since it is "atomic" with regard to the corresponding lock contention (i.e. there is no race condition going to read the lock contents when they might have changed since the lock contention).

Reviewed By: quark-zju

Differential Revision: D41998653

fbshipit-source-id: 3c80b2321a97b052025ef1f316defce3177c60ef