repolock: share locks between Python and Rust
Summary:
Hook things up so the Python repo.lock() and repo.wlock() functions delegate to the Rust repo's RepoLocker object. This way, repo locks that Python acquires are re-entrant across calls into the Rust workingcopy.

In particular, this fixes a deadlock when calling into the Rust status code since the status code needs to acquire the wlock to update the treestate.

I tweaked the out-of-order lock detection to allow the wlock to be released before the store lock if the wlock had been acquired out-of-order via a no-wait/non-blocking acquisition. The Python status code does `repo.wlock(wait=False)` in a best effort attempt to update the dirstate. The store lock may already be held, so in this case they locks are acquired out of order, and released out of order (but there is no risk of deadlock, so it is okay).

In general I've taken a conservative approach to maintain all the existing Python logic, such as acquire/release callbacks, warnings, shared wlock acquisition, etc. I considered a more aggressive change where we push more into Rust, but we can do that later.

Reviewed By: sggutier

Differential Revision: D45510005

fbshipit-source-id: 178b2213db7e6ec57c3e59c66ca3e21d27674cbf