Restart client subscriptions when the repository/cwd changes
Summary:
After the last diff, we now support changing your active repo/cwd.

This requires the client to forget most of its state—in a new repo, none of this would apply. This could have been quite difficult, but we were already basically set up to support this via serverApi.onConnectOrReconnect. A slight superset of this is the new serverApi.onSetup, which handles connect/reconnect, but also handles changes to the cwd/repo.

All our atoms which depend on server state will then automatically re-send their init message, which will then get a response pretty quickly with new info.

This seems to work well in my testing, though it's possible that for there to be some subtle bugs here.
Need to dig in a big to look for anything that might not be getting invalidated. For example, the last run command in the progress thing at the bottom does not get cleared when changing repos. Kind of a tradeoff here, because we'll lose your history of commands since it's not persisted anywhere. That's probably ok though, better have it consistent.

Reviewed By: quark-zju

Differential Revision: D45592101

fbshipit-source-id: 609a1cea64034175514764f39429cf711a491939