Add clearOnCwdChange atom effect
Summary:
Some atoms in ISL should clear their state back to the default when the repository is changed. For example, the list of previous and queued operations. If your active repo is now a different repo, those operations aren't really comparable. So it's better for us to just hide them.

Here we introduce an atom that makes it easy to declaratively reset your atom state on repo change.

Reviewed By: quark-zju

Differential Revision: D45593018

fbshipit-source-id: ad3dd0eff875d54b1083d63ab8197a88e68a8f80