Handle changing cwd/repo in isl-server
Summary:
Now that we show a dropdown selector to change your repo, start handling the changeCwd message to actually update your repo.

This requires some slight restructuring. Instead of getting the repository reference in onClientConnection, we have ServerToClientApi know how to make repo references.

This is a slight inversion of control, but since we need repository reference creation to happen in response to messages from the client, it kind of makes sense to do it in ServerToClientAPI.

Reviewed By: quark-zju

Differential Revision: D45592197

fbshipit-source-id: cdc5f6ef63e608998b0e6483cf2b4454bce5065e