Error if user passes empty --config-dir to edenfsctl
Summary:
If a user relies on an environment variable in their invocation of edenfsctl,
e.g. `--config-dir=$EDEN_DEV_STATE`, then if the user accidentally leaves the
variable unset they may inadvertently specify an empty string as the path to
their config dir.

When running `edenfsctl stop`, the effect of an empty config dir path is to
shut down your main eden daemon instead of your dev daemon, so this has proven
a bit of a footgun.

Reviewed By: xavierd

Differential Revision: D45591968

fbshipit-source-id: 6df6eb8c05d94dafcaee1d6976a52310454a42d4