Add scaffolding for tokenization and autocomplete in commit info text fields
Summary:
We want to show a typeahead/autocomplete while you're typing reviewers / subscribers / tasks/ etc. This will need to fetch data and show a list of values, then let you accept the results.

To make this nicer, we also want to show the values as tokens, instead of just text. These visual tokens distinguishes values from each other, and lets you click on the x to dismiss them more easily.

This is just the first steps in building this UX. Remaining in further commits, we need to actually fetch the right data depending on the type of field, handle the data, insert suggestions, let you delete suggestions, and deal with focus / other subtleties.

Reviewed By: quark-zju

Differential Revision: D45627742

fbshipit-source-id: d9133c04291bf6165bdf4b7efc6c22330ef68afe