Only wrap diff badges in small widths
Summary:
The old ISL showed diff badges next to commit titles, inline, with wrapping in responsive widths. This was a little annoying because things wrapped individually, and so it was scan visually since each commit was probably differently placed than the others.

We currently have been showing commit's diff information always in a second line below the commit. This makes it very consistent to find information, at the cost of vertical space. This makes ISL generally less information dense, which is not great.

Especially if you have a wide screen, this feels unnecessary.

The new approach in this diff is a sort of hybrid approach. Small screen widths will have ALL commits wrap their diff badge information onto a second line. This allows the title to wrap along side the buttons, which means buttons are in a much more consistent place on the right side, without jumping down.

On larger screens (800px), we don't wrap the diff badges, and everything is in one line. For extremely long titles, we also then allow text wrapping, but the buttons stay in place instead of wrapping.

After some testing, I actually think this system is a lot more natural feeling. Things generally stay in the place you expect them to be, and all diffs are consistent. The wrapping point at 800px is something we could adjust, but it feels about right in my testing.

One question: we could make this configurable with the old behavior, so we allow you to choose if you want to "always wrap" or "fill horizontal space".

But I think this new way is much nicer in general, so I don't think we need a config for it.

Also note that instead of using total screen width (e.g. with a CSS media query or a simple document.body.clientWidth), we actually watch the main content area and take its width instead. This allows us to be responsive respecting the commit info sidebar opening/closing,.

Reviewed By: quark-zju

Differential Revision: D45671035

fbshipit-source-id: 423b96c28b7e3fcd7faa5301d15f45b34a1ce2be