Fix json rendering for commit-info for commits with git-extra-headers
Summary: The `git_extra_headers` field in a changeset contains the headers associated with the corresponding git commit which was translated into a Mononoke changeset. The `git_extra_headers` is a mapping of arbitrary bytes to bytes which might not necessarily be valid strings. When used in `scsc`, particularly for `scsc info`, we just want to display headers that make sense to a human (i.e. strings instead of bytes). This was already the case for the default output mode but when outputting in `json`, we use `serde` serialization which doesn't accept non-string keys when serializing `BTreeMaps`. A user [faced an error](https://fb.workplace.com/groups/2071217289839797/permalink/6054134597969443/) when trying to get the info for such a commit. This diff excludes non-string headers for the purpose of displaying output to the user.

Reviewed By: YousefSalama

Differential Revision: D45689738

fbshipit-source-id: 7be727650ad7637213ae59dc36e8a73464406f1a