reduce buffer size for cbor stream from 1 MB to 8 KB
Summary:
reduce buffer size for cbor stream from 1 MB to 8 KB by default

Currently, we don’t attempt to deserialise any data less than 1 MB on cbor
stream. The value is too high for most of our requests meaning the deserialization will be
performed at the end. Since the value was chosen arbitrarily in
https://www.internalfb.com/diff/D22370164, let's adjust it to 8K. This is more
reasonable default. For example, this is the frame size BufReader is using, and so
our zstd stream.

We will still move to 1MB buffers and large once we hit resize logic if we encounter an item that doesn't fit the smaller buffer. This may happen for files endpoint.

The current implememtation doesn't allow, for example, properly track number of incoming commits coming from commit graph stream since the entries
are very small and we are blocked for long on filling the buffer.

Reviewed By: quark-zju

Differential Revision: D45619933

fbshipit-source-id: b1e9d6e97aaa67aeaa718e3127baedcd615c82f7