store: allow LocalStoreCachedBackingStore to be configured
Summary:
One of the main objective of the LocalStoreCachedBackingStore is the ability of
configuring its policy depending on the underlying backing store. For instance,
while some backing store want to cache everything in the LocalStore, others may
only want to cache BlobMetadata. Configuring the policy per-backingstore allows
the removal and special case inside the RocksDbLocalStore of blob caching: this
is now done entirely inside the LocalStoreCachedBackingStore.

Reviewed By: chadaustin

Differential Revision: D45550480

fbshipit-source-id: 386e384f0148e38fad8d4008b4e312f360edb676