auth: don't load certs if auth proxy enabled
Summary:
Skip certificate loading if the auth proxy is enabled. We don't use the certs in this case, so there is no point in loading them.

Also, when using an auth proxy (e.g. x2pagentd) the certs may not be present on disk, and we don't want to error out assuming they will be present.

Reviewed By: quark-zju

Differential Revision: D45486476

fbshipit-source-id: 3638299f72e7e5769eba48425be47aa90f96914e