config: allow BlobMetadata to not be cached in LocalStore
Summary:
One of the yet to be root caused bug happens when the BlobMetadata stored in
the LocalStore contains the wrong value. On macOS in particular, this leads to
files being truncated (or too big), but it's also possible that Buck is
receiving the wrong sha1/size when querying EdenFS.

In order to confirm that the issue is related to the LocalStore, let's add a
configuration that allows the LocalStore to not be used for BlobMetadata, which
would allow to bypass the issue described above.

Reviewed By: chadaustin

Differential Revision: D45625431

fbshipit-source-id: ffb9dbc618486e490bfa6b1e0d4391749de64bfa