tests: enable running tests on Windows using buck
Summary:
This makes some changes similar to the ones done in D45655077 for enabling tests on Windows using buck2. Hopefully this will give us more additional signals on Windows, as previously we were able to land code as our signals were not as useful.

Additionally, some unnecessary mononoke dependencies were dropped as they are not used in the `hg_run_py_tests` and `run_tests` targets

Reviewed By: muirdm

Differential Revision: D45443645

fbshipit-source-id: 559d68e442579cbe7b8dd416923c421a8dd7af42