testing: less permissive about python code blocks
Summary:
Previously, we treat code blocks that are valid Python as Python (inspired from
xonsh), otherwise comments. But that leads to issues like D45739628 where we
want the code block to be Python. It is also permissive so "dangling" output
blocks are silently ignored (see the previous change for examples).

This diff abandons the `ast.parse` approach and infers types of code blocks
strictly from indentation:
- Non-indented lines are _always_ comments, regardless of whether it passes
  `ast.parse` or not.
- Lines indented by 4 spaces, outside a `$` or `>>>` blocks are _always_ Python
  code blocks, regardless of whether it passes `ast.parse` or not.
  4-space is chosen since it matches the common Python coding practice and is
  visually and semantically different from output indentation.

Reviewed By: zzl0

Differential Revision: D45741907

fbshipit-source-id: 6d913e6f9cd5846d3c37d3c429736efaee932f5f