Don't use successor revset for hide operation
Summary:
This is a quick fix for a larger issue that needs better addressing.

Currently, if you tried to hide a commit via right click, it would track successors and hide the latest descendant. This is good for most operations, but definitely not hide. A common use case for hiding is to remove an obsolete commit from your tree. Usually this happens automatically, but if it doesn't or your commit is shown for some other reason, it's common to want to hide.

If this used the successor revset, it would NOT hide the obsolete commit and instead hide a perfectly good draft commit!

What we should really do for this is probably to check if the commit you ask to hide is obsolete, and if so don't use the successor revset. If not, it could potentially use the successor. But this gets complicated with queued commands and optimistic state. We may need to enforce that optimistic commits don't get hidden or something.

I think for now we can just make hiding an extremely explicit operation that doesn't track successors at all.

Reviewed By: quark-zju

Differential Revision: D45750878

fbshipit-source-id: da5bba9e2e2911c492a7cbc0e9106f2774810e88