Disable buttons that run operations automatically while running
Summary:
Make a wrapper component, `OperationDisabledButton` for the specific but common use case of a button that runs an Operation and wants to be disabled while its running.

Previously, we weren't disabling the submit/amend/commit buttons while those commands ran. That's not strictly wrong, because of our queueing system. But it's pretty useless to run submit twice in a row.

The idea is that we only look at at most one current pending operation—the topmost queued command or if none are queued, the current running operation as long as it hasn't exited yet.

This means queueing up any other command while the submit/commit/amend are still running will "undisable" the button. This reflects the possibility that the operation(s) you ran may make it reasonable to want to queue a new commit/amend/submit. In practice, this probably doesn't matter. But it is much cheaper for us to only check one operation to match, and makes this component very easy to use everywhere we run commands.

The component itself converts `onClick` into a function which is expected to return an Operation (or multiple, or a promise of operation(s)). It then runs the operation on your behalf, and also remembers which operation it is so it knows if it's still the current pending operation.

You just provide a "contextKey" to remember this run operation with. This allows the same button in different contexts to not get disabled together, for example submitting from the commit info view

Reviewed By: quark-zju

Differential Revision: D45716179

fbshipit-source-id: 3dc88634625a26e0bf98d63dc24bb84b8ab4df91