remotefilelog: don't sort file ancestors()
Summary:
Avoid the linkrev sort of remotefilectx.ancestors() by default. If there is a file node with a linknode not present in the nodemap, it triggers a full changelog scan looking for the commit with the filenode.

It seems like the sorting was needed for legacy copy tracing (in particular, edenscm.copies._related assumes the ancestors are sorted). However, with the new copy tracing, this code is dead or at least almost dead. I left in a compatibiltiy sort=True option just in case.

Reviewed By: sggutier

Differential Revision: D45756773

fbshipit-source-id: b190cc8e0b5bd0974b4ceec49d293df965f75c46