Collect logs from VS Code extension in memory
Summary:
VS Code's output channel API doesn't give you a way to get data out of the channel. So if we want to attach your logs to a bug report, we have no way to do that.

Note that this is only used internally, OSS has no file-a-bug system (we just use issues).

Reviewed By: quark-zju

Differential Revision: D45739765

fbshipit-source-id: a58d350f274b60928d329b9f76d3784cead1c50b