Show "public" badge in commit info view
Summary:
Public commits are discouraged from being shown in the commit info. But it is still possible. We should make it more obvious that you're looking at a public commit. In later diffs this stack we'll also make them read-only and hide submit/amend buttons.

Add an InlineBadge component to support "You are here"-like badges in other places in the UI.

Reviewed By: quark-zju

Differential Revision: D45742163

fbshipit-source-id: 18ade6bba15349b86691e7700092e01bee52055f