use more meaningful progress bar
Summary:
use more meaningful progress bar for commit graph eden api endpoint

If we show just speed here, it maybe confusing in edge cases.

For example, for AOSP repo, we may keep saying that we download just few KB for 10 and more minutes.

If we say, we are fetching changesets, this mean we calculate things server side, so less confusing. It would be probably even better not to show the downloading progress bar but not sure how to disable it.

Reviewed By: muirdm

Differential Revision: D45690720

fbshipit-source-id: efdc193ba6f15961cbf2bdaf8b00dff0700c9c59