disable response compression
Summary:
all responses in edenapi are zstd compressed on _all_ platforms.

this is enabled here:

https://www.internalfb.com/code/configerator/[645398df35269eed7047d9d1c39015c39c038d26]/source/scm/hg/hgclientconf/hgclient.cinc?lines=NaN-NaN%2C1031%2C1033%2C1035

This isn't really needed for many many of our responses since they are anyway
very small. On the other hand, it is probably not a big deal but for commit
graph endpoint I have implemented a special progress bar that updates much nicer if we disable
such compression. So, I would like to disable it explicitly.

Reviewed By: muirdm

Differential Revision: D45691469

fbshipit-source-id: 67c3debe473638dfffc96084b7ad438870a26951