runlog: silence error output
Summary:
The user doesn't really care if a runlog update failed. This is expected to happen sometimes on Windows since the update will fail if the file is underlying runlog file is open (by another process).

Also, add a couple retries to the final runlog update on Windows to increase the chance we close out the entry.

Reviewed By: sggutier

Differential Revision: D45831199

fbshipit-source-id: 664e14695fdf3a7363dafc7523cef78c9fafd84c