debugstack: add a way to export extra files in `debugexportstack -r wdir()`
Summary:
By default, `debugexportstack -r wdir()` exports tracked file changes. In the
ISL use-case, we'd like to add a "chunk selector" for files in the working copy.
The "chunk selector" would like to know the file contents regardless of whether
files are in `?`, `M`, `A`, `!`, `R` states.

This diff adds a `--assume-tracked` flag to support this use-case. Basically,
this flag can be used to make files in `!` or `?` states exported as if they
were `R` and `A` states.

Reviewed By: muirdm

Differential Revision: D45787979

fbshipit-source-id: 812a389ffe6d01b8a4200754bdc38c49b775c67e