Implement endpoint for creating tagged changeset
Summary: This diff implements the `create_git_tag` endpoint introduced in the last diff. The primary logic is implemeted as part of the `mononoke_api::create_annotated_tag` method so no new logic is introduced here. The `create_annotated_tag` does not yet store the mapping of the created changeset object with the tag name

Reviewed By: mitrandir77

Differential Revision: D45654021

fbshipit-source-id: 3ebd8d7b67a59935f646f5ba4287e14a096f82d1