introduce EdenMountHandle
Summary:
Asynchronous operations that must ensure an EdenMount (and its root)
are usable for the duration must hold a reference to the root's
InodePtr, even if they only need the EdenMount.

To make this relationship less implicit, introduce an EdenMountHandle
which is effectively the same as the previous
`pair<EdenMountPtr, TreeInodePtr>`, but more clearly documents the intent
is to preserve the usability of EdenMount and its root inode.

This also allows us to decouple that relationship in the future,
should we want to.

For now, only use EdenMountHandle in getMountPoints. Next, I'll switch
`getMount` to use it.

Reviewed By: kmancini

Differential Revision: D45830176

fbshipit-source-id: 832dd96b9a6fe04fabbb5af0ecbed0173f9d2a2b