Modify ObjectCacheBench to test both long and short IDs
Summary:
We are considering changing the ObjectID hash function to no-longer be constant time. Instead, it might iterate over the bytes of the ID and "mix-in" contents from each 8 byte chunk into a final hash.

In order to measure the performance hit this might have, we need to benchmark both long ObjectIds and short ObjectIDs. This diff introduces a benchmark to test hashing ObjectIDs of length 240 characters long, which is a realistic size of our current ObjectIDs.

Reviewed By: chadaustin

Differential Revision: D45798645

fbshipit-source-id: 988442555f48968edf67517aa25b55ef636de610