tracing: simplify tracing collector
Summary:
The tracing collector receives tracing spans/events and sticks them in an in-memory TracingData structure. This data is used for to print out the perftrace, among other things.

Previously the collector implemented both Layer and Subscriber. However, Subscriber is not convenient since it cannot be composed with Filters. In this commit, I get rid of the Subscriber so we are only dealing with a Layer. I also got rid of the built-in LevelFilter to simplify things. It is up to the user of the TracingCollector to apply the appropriate Filter. (With no filter applied, everything will be collected.)

Reviewed By: quark-zju

Differential Revision: D45922413

fbshipit-source-id: ea1bf03ca56bc11f66e05d4298fb0bd638a650d2