Implement BonsaiTagMapping in SQL
Summary: This diff implements the SQL logic for creating and fetching entries from `bonsai_tag_mapping` table. It also defines a new facet which will be used in gitimport, remote-gitimport and mononoke-admin tool later on. I will send a follow-up diff with tests to validate the behavior of the SQL implementation.

Reviewed By: mitrandir77

Differential Revision: D45904551

fbshipit-source-id: 31a99d89b10b39ce7aeda5f4e46b0c94ec6b60ba