Create bonsai_tag_mapping table
Summary:
When importing a Git commit into Mononoke, we store the tag as a bookmark and the commit that the tag points to becomes a changeset in Mononoke. The bookmark then points to the newly created changeset. However there is currently no mapping from the tag to the changeset that contains the tag metadata. This table has been created for that purpose.

This diff lays the foundation for the `facet` `BonsaiTagMapping` which will be implemented in follow-up diffs.

Reviewed By: mitrandir77

Differential Revision: D45867181

fbshipit-source-id: 1c251e40b081eb38f217f5f7c85095c574263b24