test for repo working copy verification
Summary: During the change of mapping between repos we have to check if the new mapping applies cleanly to current repo  mainline bookmark. To do that so far we've been using a binary that compared content ids for all relevant file paths. I want to speed it up but before I can comfortably do it I need to extend the test coverage. We had very little test coverage which covered only. happy paths and simplest configs.

Reviewed By: liubov-dmitrieva

Differential Revision: D45864219

fbshipit-source-id: edd78e88796e65268aa549bb443421506fe1507e