commit_graph: implement common_base
Summary: common_base returns all of the highest generation changesets that are common ancestors of two changesets. Implementation is optimized for the most common case were there are no merges between the two changesets and their lowest common ancestor (taking O(log(height) iterations in that case), but is sound for all cases.

Reviewed By: mitrandir77

Differential Revision: D45480232

fbshipit-source-id: 147b6feaa57cdbeb5efb4b77335728f38105fb87