Utilize ContentMetadataV2 in metadata-diff endpoint
Summary:
The `metadata-diff` endpoint in SCS is used by Phabricator to get information about the difference between files of two commits. Currently, this endpoint works using raw file content. This diff makes the following optimizations:
- Avoids loading file content in memory if the file is `binary` or `non-utf8`
- Avoids parsing `generated` and `partially-generated` tags if the file is known to be neither
- Avoid parsing raw file bytes into UTF-8 string if the file is non-generated

These optimizations will be most helpful for the case of `binary` or `non-utf8` files but should improve performance of the endpoint in general.

Reviewed By: YousefSalama

Differential Revision: D46109336

fbshipit-source-id: b3c915005686f77defceeb53f1587049bc8b6bd8