Modify bookmark get admin command to handle annotated tags
Summary: The bookmark `get` command simply displays the changeset that is pointed to by the bookmark. This diff modifies the command to also handle git annotated tags converted into bookmarks. If the bookmark is a tag bookmark, then the command also outputs the metadata changeset associated with the bookmark. This will be useful for validating that imported annotated tags retain their metadata through metadata changesets.

Differential Revision: D46022692

fbshipit-source-id: b3980f4799066913268ef7acc1f4956d3feb5b23