remove debugGetScmBlobMetadata
Summary:
I migrated the last caller of debugGetScmBlobMetadata to debugGetBlobMetadata
in the last diff. Let's delete the old method.

I am not really concerned about waiting for that lower diff to rollout and all
because this is a debugging command that we only sparingly ask users to run
(As far as I know like maybe 3 times).

Reviewed By: chadaustin

Differential Revision: D45972617

fbshipit-source-id: 6d4689a225fee972f8e3a802115852043f0bade8