Add better loading state for pull request viewing
Updates the PR page to show the spinner while blobs/diffs are loading. Currently, the Suspense wrapping this is never triggered, leading to two big issues:
1. The page appears empty while loading (there is no indication of loading)
2. As the diffs load, they will appear in random order, causing layout shift until all diffs have loaded.

The only concern of this PR is that every diff must load to see anything now, but it was unusable before anyway as the layout would shift until all diffs were loaded anyway.